Store the GStreamer registry/cache in our own cache directory. Fixes #2839
This works around the cache bug linked in #2839 by forcing new version to
create a new cache. In case there are more problems in the future we might
want to delete the cache when we update.

With our own cache we also avoid fighting over the cache with other GStreamer
using applications, or when switching between a msys2 dev build and the
system installed one.