mpris: write embedded covers into the cache dir and not /tmp. Fixes #2847
In case we get a file in /tmp copy it to the cache directory. This turned out to
be a bit problematic because we call that function quite often and /tmp usually
being a tmpfs probably helped. Try to improve the situation by not fetching
the metadata every time the property is fetched but cache it and invalidate
it in the right places.