wininst: fix install window not being in front when uninstall finishes
A bit hacky, but the only way I found to make this work.
Move the uninstall code into onGUIInit and call BringToFront so we
create a semi finished install window before the uninstall window opens.

This results in the install window being in focus when the uninstall part
is finished.