Fix setting pane mode
When setting `pane_mode` via the GUI, the config file contains something
like this:
```
pane_mode = ColumnMode.COLUMNAR
```

This is due to the fact that ColumnMode is an enum and casting it to a
string returns the enum name. However, the config code expects an int.
Manually setting `pane_mode` to an int in the config file fixes the
problem, so we just cast it to an int before setting the config value.