Don't clear queue when set_explicit returns None
When a new index in a model is set with go_to,
order.set_explicit() is called. But this function is allowed to
return None, and documentation says that "no action will be taken
by the player."

This commit fixes an issue where an action *is* taken: calling
PlaylistMux's go_to method will unconditionally reset the model it
thinks is inactive.

This surfaces as a user visible issue when using persistent queue
mode and some extension (like Queue Only) that expects to be able
to override set_explicit(). When activating an item in a song list,
go_to is called with the song list as a source, so Queue Only will
fail to prevent the queue from losing its state.

Fixes #3385