Improve the console plug-in in a number of ways (#3923)
* allow using tab key for autocomplete in console ext

* Smarter autocomplete in console ext

If there is only one suggestion, it is filled automatically. If
the suggestion would do nothing (because the text already exists
on the prompt), add a '.' instead to allow accessing properties,
if any properties are currently visible.

* Fix colorizing overscroll in console ext

The wildcard CSS resulted in a weird looking over/under scroll
element being styled white. Fix that and make the color use the
user's theme for backgrounds, if possible.

* console ext: return blank list when no autosuggestion

As part of other changes, we now check len() to determine
the correct behavior, so we need an empty list when no
suggestions are available.

* console ext: add handling for ctrl-c

Deletes the current (possibly multi-line) entry if ctrl-c is pressed.
Also improves handling around automatic indentation for <return> key.