Strip blank tags out from metadata lists (#3914)
Remove empty tags from lists of tags returned from AudioFiles.
Also applies to formatted lists of tags (comma separated) and
lists with pairs of display and sort tags ("tag", "tagsort").

Reasoning:

It's common to have a tag in a file be present but blank. When
this happens, lists of metadata used in collections (e.g. by the
~persons tag) will contain empty strings and be combined and exposed
to the user in ugly ways. (Example: "Artist 1, , Artist 2,")

Previously, this behavior was only present for synthetic tags, but
these changes apply them to real tags as well.

Fixes #3383