Allow auto-playing added songs in Queue Only mode (#3947)
When Queue Only mode is enabled, the user has multiple
neutral methods to add a song to the queue: both
Ctrl+Enter and the right click menu. Double clicking a
song usually indicates intent to play it immediately.

Queue Only mode changes the double click action to add
the song to the queue instead, but some users might like
Quod Libet to respect their intent and jump to the newly
added queue item and begin playing it. This adds the
ability to do that to the Queue Only plugin, and puts it
behind a preference.

Fixes #3381.

Implementation notes:

The code for jumping in the queue is specifically designed
to jump to the first added item in the queue. This is to
make implementing a double click action for albums easier
in the future. (QL will need to jump to the frist song in
the added album.)

A change had to be made to the player code, because it was
emitting an explicit go_to when the main window restores
the last playing song when starting up and this was getting
caught by the plugin. Intuitively, restoring a previously
playing song is an implicit action by the player.