Clip images for rounded border (#4118)
The function `add_border` now first creates a wide path along the
edge of the final image to draw the border, then it creates a narrower
path in which to draw the original image. This way the border-radius is
actually applied to the original image, which was not the case before.

Fixes #4117.