Improve efficiency of tag pattern cache
Currently the code clears che cache and starts over whenever it
reaches its capacity. This change uses an OrderedDict in place of
the existing dict based cache. It behaves the same, but guarantees
insertion ordering on older platforms, so we can use the popitem
method to remove old cache entries as necessary.

We also use the move_to_end method to cheaply keep recently
accessed items at the front of the cache, where they are not at
risk of expiry.

Benchmarking with a set of 200 simple tags and assuming random
access shows a performance improvement of 30%.

In the worst case scenario for the original approach, the cache
always expires an entry right before it is needed. In this case
this commit is about 90 *times* faster.

In no scenario I tested, including the case where the cache is
never cleared because fewer than `MAX_CACHE_SIZE` tags patterns
are used, was the new code significantly slower than the old code.