Disable ordering when active order disappears
When you disable a plugin-based play order or repeat mode, the UI
is refreshed so that it will stop showing the mode in the menu.
However, this change is only visual. The mode remains active until
the app is restarted or the user actively changes the mode.

I considered several ways of fixing this, but I ended up opting
to just disable the play mode entirely until the user takes action.
This is the only thing that makes sense, because a user who has
been using "queue only" mode probably doesn't want to be switched
to "shuffle" mode just because they decide to disable the plugin.

Fixes #4155.